backend/requirements.txt
Flask
flask-cors
requests
scikit-learn
numpy
pandas
python-dotenv
